treewide: Refactor SHA256 driver into an implemented trait
`Sha256HardwareDriver` is now the primary implementation of
`Sha256`. Where a concrete implementation is required,
`Sha256HardwareDriver` is almost always the correct instance.

This refactor makes it easier to test consumers of this driver, where
software implementations of SHA256 may be preferred. Follow-up commits
will provide a software implementation, which will then be used for
fuzz testing the LMS driver.

Signed-off-by: Benjamin Doron <benjamin.doron@9elements.com>